wip: refactor before implementing tree editing
